SCP-4636 Generalized detection, discovery, and processing of contracts.
